feat: Soft Delete cho những dữ liệu quan trọng, tránh force delete

+++ BrainEduBackend/src/main/java/com/brainedu/BrainEdu/service/instructorService/InstructorAssignmentServiceImpl.java
@@ -73,7 +73,7 @@ public class InstructorAssignmentServiceImpl
                                 request.getDueDate()
                         )
                         .maxScore(
-                                request.getMaxScore()
+                                request.getMaxScore().floatValue()
                         )
                         .createdAt(
                                 LocalDateTime.now()
